requirements.txt
click==8.1.1
Flask==2.1.1
Flask-Cors==3.0.10
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
numpy==1.22.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
Werkzeug==2.1.0
zipp==3.7.0
